SCUMM: Add a simple 8253/8254 Programmable Interval Timer handler
In V5–V7, the PIT frequency does not correspond to the game speed anymore, though the speed _is_ derived from it in a strange way.

Since the SCUMM interpreters replace the DOS timer interrupt handler with their own code, the system clock will still need to be updated at its normal rate or it will go out of sync. It is our misfortune that LucasArts did everything so carefully here (unlike other engines), as it makes it difficult to replicate the timing of the V5–V7 games.

The 236.6955 Hz frequency from the previous games is related to their handling of this. This is roughly 13 times the base PIT frequency of 18.2065 Hz and one can trivially derive from it the interrupt rate that the system clock will be expecting.

However, in V5–V7 LucasArts went slightly into over-engineering territory, to make the PIT frequency an exact multiple of 18.2065 Hz—291.3042 Hz. To achieve the 240 Hz frequency that we need for the wait loop, they simply skipped every nth interrupt—the same trick that is normally used to keep the system clock in sync.

I do not fully understand the code in the interrupt handler that deals with the 291.3042 Hz to 240 Hz conversion, and adding to the confusion is that unlike the preceding games, V5 games run at very close to 60 frames per second, but V6 games run significantly _slower_ than 60 frames per second and even a bit slower than the 236.6955 Hz frequency of the preceding games.

I manually profiled the speed in V6 for now. This is accurate to one decimal place maybe.