SCUMM: Implement scrollEffect() properly
One would expect the velocity of scrolling—not the duration—to be constant. Indeed, this is what the original interpreter did.

This implementation is bugged for another reason though, as the mangled quarter-frame to millisecond conversion (which should be fractional) causes the step calculation not to round evenly. The horizontal step length is unexpectedly truncated.

Not to mention, if we _would_ convert VAR_FADE_DELAY correctly, the result would get mangled anyway when passed to waitForTimer(), which causes a drift of one millisecond per two frames; roughly one frame per transition.

Finally, and most notably, the lack of sleep granularity in waitForTimer() effectively turns the 12 millisecond delays per step into 20 millisecond delays, causing a terribly slowed screen transition visible to the naked eye.