SCUMM: Implement fractional time handling
kFadeDelay is a constant that deals with the quarter-frame to millisecond conversion.

We should not convert anything here, though. Instead, we should use quarter-frames as the base unit of time as the original interpreter did, only converting once—carefully—in waitForTimer().

While these internal loops use quarter-frames for timing, and the main wait loop uses whole frames, in neither case is it okay to truncate the result from the conversion. It will cause clock drift.