SAGA: Remove spurious CP 850 encoding
Why are we mapping the in-game fonts like this? The fonts are what you would expect to be mapping _to_. In fact, this behavior seems to have been the cause of several bugs over the years.

Adding to the mystery, the DOS fonts are actually Windows-1252. Well, this makes sense in a way because we are _decoding_ from CP 850 to ISO-8859-1, but then where is this CP 850 encoding coming from to begin with?

We should just map the hardcoded string literals to Windows-1252, and we are done.