BACKENDS: Amend KeyCode fallback, avoid relying on the Num Lock state
We cannot keep track of the Num Lock state on all platforms—bugs #10558 and #11227, nor can we rely on SDL to produce the KP keysyms on all platforms. They are dependent on the Num Lock state. At least on Windows we will get arrow key events instead of the KP2, KP4, KP6 and KP8 keysym events, when Num Lock off.

These hardcoded mappings are ultimately obsoleted by the Common DOS mapping for the relevant engines, though that does not solve the Num Lock issues per se.